build: create a pypi cli tool
